Fix tests
The behavior has changed with https://github.com/codehaus-plexus/plexus-utils/issues/216 but plexus-utils contains duplicate merge code in Xpp3Dom and Xpp3DomUtils and the two pieces have divered.  This duplication is removed by the PR by switching to the XmlNode implementation in both cases, but we now need to fix the expectations.